backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.31
asyncpg==0.30.0
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.3.2
pydantic==2.7.4
pydantic-settings==2.3.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
aiofiles==23.2.1
redis==5.0.6
pdfplumber==0.11.1
python-docx==1.1.2
chardet==5.2.0
httpx>=0.25.2
pytesseract==0.3.10
Pillow==10.4.0
numpy==1.26.4

# ── LangGraph + LangChain ──────────────────────────────────────────────────
langgraph==0.2.28
langchain==0.3.4
langchain-core>=0.3.15
langchain-groq==0.2.1

# ── Локальные эмбеддинги (sentence-transformers, без API) ────────────────
sentence-transformers==3.3.1
torch==2.4.1

# ── Telegram Bot ───────────────────────────────────────────────────────────
python-telegram-bot==21.4
python-dotenv==1.0.0